working on better skipping of unnecessary stage1 transforms
This adds filtering for a newer stock plugin added by ember-cli-htmlbars. Next we need to also do this filtering before counting the present plugins in `needsCustomBabel` to skip it entirely.
