+++ packages/compat/src/detect-babel-plugins.ts
@@ -1,0 +1,37 @@
+import { PluginItem } from '@babel/core';
+
+export function isEmberAutoImportDynamic(item: PluginItem): boolean {
+  let pluginPath: string;
+  if (typeof item === 'string') {
+    pluginPath = item;
+  } else if (Array.isArray(item) && item.length > 0 && typeof item[0] === 'string') {
+    pluginPath = item[0];
+  } else {
+    return false;
+  }
+  return /(^|\/)ember-auto-import\//.test(pluginPath);
+}
+
+export function isCompactReexports(item: PluginItem): boolean {
+  let pluginPath: string;
+  if (typeof item === 'string') {
+    pluginPath = item;
+  } else if (Array.isArray(item) && item.length > 0 && typeof item[0] === 'string') {
+    pluginPath = item[0];
+  } else {
+    return false;
+  }
+  return /(^|\/)babel-plugin-compact-reexports\//.test(pluginPath);
+}
+
+export function isColocationPlugin(item: PluginItem): boolean {
+  let pluginPath: string;
+  if (typeof item === 'string') {
+    pluginPath = item;
+  } else if (Array.isArray(item) && item.length > 0 && typeof item[0] === 'string') {
+    pluginPath = item[0];
+  } else {
+    return false;
+  }
+  return /(^|\/)ember-cli-htmlbars\/lib\/colocated-babel-plugin/.test(pluginPath);
+}
